feat(app): user와 computer class를 각각의 파일에 분리하여 구현
